311_backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
kafka-python
pandas
sodapy
pyspark
